% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey_api.R
\docType{data}
\name{app}
\alias{app}
\title{OAuth application for SurveyMonkey.}
\format{An object of class \code{oauth_app} of length 4.}
\usage{
app
}
\description{
See \code{\link[httr:oauth_app]{httr::oauth_app()}} for clarification of arguments.
}
\keyword{internal}
