% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey_api.R
\docType{data}
\name{.state}
\alias{.state}
\title{Environment to store credentials in.}
\format{An object of class \code{environment} of length 0.}
\usage{
.state
}
\description{
Environment to store credentials in.
}
\keyword{internal}
