% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_surveys.R
\name{list_surveys}
\alias{list_surveys}
\title{Return a table of surveys owned or shared by the user.}
\usage{
list_surveys()
}
\value{
A tibble listing all the accessible surveys with the following columns:
\itemize{
\item \code{id} <chr>: Survey identifer.
\item \code{title} <chr>: Survey name.
\item \code{href} <chr>: - Survey URL.
\item \code{nickname} <chr>: Survey nickname.
}
}
\description{
Return a table of surveys owned or shared by the user.
}
\examples{
\dontrun{
list_surveys()
}
}
