% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_responses.R
\name{matrix_ranking}
\alias{matrix_ranking}
\title{matrix_ranking}
\usage{
matrix_ranking(x)
}
\arguments{
\item{x}{A named list with \code{responses}, \code{choices} and \code{rows} tibbles.}
}
\value{
A tibble with a single row per response and the following structure:
\itemize{
\item \code{response_id} <chr>: Id for the response.
\item \code{<row name>} <int>: Ranking given for the row. One column per question row.
}
}
\description{
matrix_ranking
}
