% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_responses.R
\name{matrix_rating}
\alias{matrix_rating}
\title{Parse a list of arguments into a \code{matrix_rating} response.}
\usage{
matrix_rating(x)
}
\arguments{
\item{x}{A named list with \code{responses}, \code{choices} and \code{rows} tibbles.}
}
\value{
A tibble with a single row per response and the following structure:
\itemize{
\item \code{response_id} <chr>: Id for the response.
\item \code{<row name>} <chr>: (If question is a matrix of choices) Text entered for the row. One column per question row.
\item \code{weight>} <int>: (If question is a rating) Rating selected for the response.
}
}
\description{
Parse a list of arguments into a \code{matrix_rating} response.
}
\keyword{internal}
