% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_responses.R
\name{multiple_choice}
\alias{multiple_choice}
\title{multiple_choice}
\usage{
multiple_choice(x)
}
\arguments{
\item{x}{A named list with \code{responses}, \code{choices} and \code{rows} tibbles.}
}
\value{
A tibble with a single row per response and the following structure:
\itemize{
\item \code{response_id} <chr>:  Id for the response.
\item \code{<choice name>} <lgl>:  Logical indicating if the choice was selected. One column per choice.
\item \code{text_other} <chr>:  (Optional) Text given an "Other" choice if it exists.
}
}
\description{
multiple_choice
}
