% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_responses.R
\name{open_ended_single}
\alias{open_ended_single}
\title{Parse a list of arguments into a \code{open_ended_single} response.}
\usage{
open_ended_single(x)
}
\arguments{
\item{x}{A named list with \code{responses}, \code{choices} and \code{rows} tibbles.}
}
\value{
A tibble with a single row per response and the following structure:
\itemize{
\item \code{response_id} <chr>: Id for the response.
\item \code{text} <chr>: Text response.
}
}
\description{
Parse a list of arguments into a \code{open_ended_single} response.
}
\keyword{internal}
