% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_api.R
\name{print.sm_api}
\alias{print.sm_api}
\title{Print a \link[=sm_api]{sm_api} response object.}
\usage{
\method{print}{sm_api}(x, ...)
}
\arguments{
\item{x}{An \link[=sm_api]{sm_api} response object.}

\item{...}{Extra arguments needed for method reasons.}
}
\value{
The \link[=sm_api]{sm_api}, invisibly.
}
\description{
Print a \link[=sm_api]{sm_api} response object.
}
\examples{
\dontrun{
surveys_raw <- sm_api("v3/surveys")
surveys_raw
}
}
