% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_responses.R
\name{single_choice}
\alias{single_choice}
\title{single_choice}
\usage{
single_choice(x)
}
\arguments{
\item{x}{A named list with \code{responses}, \code{choices} and \code{rows} tibbles.}
}
\value{
A tibble with a single row per response and the following structure:
\itemize{
\item \code{response_id} <chr>:  Id for the response.
\item \code{text} <chr>:  Text of the option chosen.
}
}
\description{
single_choice
}
