% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{sm_surveys}
\alias{sm_surveys}
\title{List your surveys}
\usage{
sm_surveys()
}
\value{
A \code{tibble} of surveys with the following columns:
\itemize{
\item \code{id} - survey id.
\item \code{title} - survey title.
\item \code{nickname} - survey nickname.
\item \code{href} - a URL to the survey.
}
}
\description{
List your surveys
}
\examples{
\dontrun{
sm_surveys()
}
}
