% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_responses.R
\name{survey_questions}
\alias{survey_questions}
\title{survey_questions}
\usage{
survey_questions(id)
}
\arguments{
\item{id}{An id for the survey. Can be found using \code{\link[=list_surveys]{list_surveys()}}.}
}
\value{
A tibble of the questions and their details. The following columns are included:
\itemize{
\item \code{question_id} <chr>:  The id of the question.
\item \code{heading} <chr>:  The question name.
\item \code{family} <chr>:  The question family and subtype that define the question type.
\item \code{subtype} <chr>:  The question family and subtype that define the question type.
\item \code{choices} <list>:  A table of choice ids and attributes.
\item \code{rows} <list>:  A table of row ids and attributes.
}
}
\description{
survey_questions
}
\examples{
\dontrun{
survey_questions(12345679)
}
}
