% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey_api.R
\docType{data}
\name{surveymonkey}
\alias{surveymonkey}
\title{OAuth endpoints for SurveyMonkey.}
\format{An object of class \code{oauth_endpoint} of length 3.}
\usage{
surveymonkey
}
\description{
See \code{\link[httr:oauth_endpoint]{httr::oauth_endpoint()}} and
\href{https://developer.surveymonkey.com/api/v3/#oauth-2-0-flow}{SurveyMonkey API developer docs}
for more details.
}
\keyword{internal}
