% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey_api.R
\name{surveymonkey_api}
\alias{surveymonkey_api}
\title{Make GET requests to the SurveyMonkey API.}
\usage{
surveymonkey_api(path)
}
\arguments{
\item{path}{URL path for the SurveyMonkey API.}
}
\value{
List representing the JSON response.
}
\description{
Make GET requests to the SurveyMonkey API.
}
\examples{
\dontrun{
surveymonkey_api("v3/surveys")
}
}
