% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey_api.R
\name{surveymonkey_token}
\alias{surveymonkey_token}
\title{Return a SurveyMonkey OAuth2 token.}
\usage{
surveymonkey_token()
}
\value{
SurveyMonkey OAuth2 token.
}
\description{
Return a SurveyMonkey OAuth2 token.
}
\keyword{internal}
