% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_surveys.R
\name{list_surveys}
\alias{list_surveys}
\title{list_surveys}
\usage{
list_surveys()
}
\value{
A tibble listing all the accessible surveys with the following columns:
\itemize{
  \item id - Survey identifer.
  \item title - Survey name.
  \item href - Survey URL.
  \item nickname - Survey nickname.
}
}
\description{
list_surveys
}
\examples{
\dontrun{
list_surveys()
}
}
