% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_responses.R
\name{open_ended_multi}
\alias{open_ended_multi}
\title{open_ended_multi}
\usage{
open_ended_multi(x)
}
\arguments{
\item{x}{A named list with \code{responses}, \code{choices} and \code{rows} tibbles.}
}
\value{
A tibble with a single row per response and the following structure:
\itemize{
\item \code{response_id} <chr>: Id for the response.
\item \code{<row name>} <chr>: Text entered for the row. One column per row.
}
}
\description{
open_ended_multi
}
