% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_responses.R
\name{survey_questions}
\alias{survey_questions}
\title{Return a table of questions for a survey.}
\usage{
survey_questions(id)
}
\arguments{
\item{id}{An id for the survey. Can be found using \code{\link[=list_surveys]{list_surveys()}}.}
}
\value{
A tibble of the questions and their details. The following columns are included:
\itemize{
\item \code{question_id} <chr>:  The id of the question.
\item \code{heading} <chr>:  The question name.
\item \code{family} <chr>:  The question family and subtype that define the question type.
\item \code{subtype} <chr>:  The question family and subtype that define the question type.
\item \code{choices} <list>:  A table of choice ids and attributes.
\item \code{rows} <list>:  A table of row ids and attributes.
}
}
\description{
Return a tibble of questions for a survey owned or shared by the user. There
is a \code{question_id} used to identify the question and a \code{heading} for
the question name, as well \code{family} and \code{subtype} which identify the type of
question. \code{choices} and \code{rows} are two fields (optionally) found in questions
and and have their own \code{id}s identifying them as well as attributes such as
\code{text} specifying what the survey taker sees.
}
\examples{
\dontrun{
survey_questions(12345679)
}
}
\keyword{internal}
