% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_responses.R
\name{survey_responses}
\alias{survey_responses}
\title{survey_responses}
\usage{
survey_responses(id)
}
\arguments{
\item{id}{An id for the survey. Can be found using \code{\link{list_surveys}}.}
}
\value{
A wide tibble of responses.
}
\description{
survey_responses
}
\examples{
\dontrun{
survey_responses(123456789)
}
}
